+++ lib/mix/test/mix/tasks/test_test.exs
@@ -32,9 +32,14 @@ defmodule Mix.Tasks.TestTest do
       assert ex_unit_opts_from_given(formatter: "A.B") == [formatters: [A.B]]
     end
 
+    test "accepts custom :exit_status" do
+      assert {:exit_status, 5} in ex_unit_opts(exit_status: 5)
+    end
+
     test "includes some default options" do
       assert ex_unit_opts([]) == [
                autorun: false,
+               exit_status: 2,
                failures_manifest_file:
                  Path.join(Mix.Project.manifest_path(), ".mix_test_failures")
              ]
@@ -48,7 +53,7 @@ defmodule Mix.Tasks.TestTest do
     defp ex_unit_opts_from_given(passed) do
       passed
       |> ex_unit_opts()
-      |> Keyword.drop([:failures_manifest_file, :autorun])
+      |> Keyword.drop([:failures_manifest_file, :autorun, :exit_status])
     end
   end
 
@@ -465,6 +470,18 @@ defmodule Mix.Tasks.TestTest do
         refute output =~ "Test suite aborted after successful execution"
         output = mix(["test", "--failed"])
         assert output =~ "2 failures"
+      end)
+    end
+  end
+
+  describe "--exit-status" do
+    @describetag :unix
+
+    test "returns custom exit status" do
+      in_fixture("test_failed", fn ->
+        {output, exit_status} = mix_code(["test", "--exit-status", "5"])
+        assert output =~ "2 failures"
+        assert exit_status == 5
       end)
     end
   end